refactor: from Arrow to Result4k
Moves the error handling logic to Result4k constructs. Arrow is more
complicated than I'd like, and is definitely overkill for my usage. Some
methods were also refactored to not return erros at all, and use
exceptions instead. This was done following the principle of only
returning errors that can be handled in the domain layer, using
exceptions for everything else. This approach was heavily inspired by
this blog post:
https://oneeyedmen.com/failure-is-not-an-option-part-7.html